+++ bothub_cli/lib.py
@@ -8,7 +8,8 @@ import json
 import time
 import traceback
 
-from six.moves import input
+from prompt_toolkit import prompt
+from prompt_toolkit.history import FileHistory
 
 from bothub_client.clients import NluClientFactory
 
@@ -26,8 +27,6 @@ from bothub_cli.utils import extract_dist_package
 from bothub_cli.utils import make_event
 from bothub_cli.utils import tabulate_dict
 from bothub_cli.utils import get_bot_class
-from bothub_cli.utils import load_readline
-from bothub_cli.utils import close_readline
 
 
 class Cli(object):
@@ -175,25 +174,23 @@ class Cli(object):
 
     def test(self):
         self._load_auth()
-        load_readline()
+        history = FileHistory('.history')
 
         project_id = self._get_current_project_id()
         bot = self._load_bot()
 
-        line = input('BotHub> ')
-        while line:
+        while True:
             try:
+                line = prompt('BotHub> ', history=history)
+                if not line:
+                    continue
                 event = make_event(line)
                 context = {}
                 bot.handle_message(event, context)
-                line = input('BotHub> ')
-            except EOFError:
+            except (EOFError, KeyboardInterrupt):
                 break
             except Exception:
                 traceback.print_exc()
-                line = input('BotHub> ')
-
-        close_readline()
 
     def add_nlu(self, nlu, credentials):
         self._load_auth()